requirements.txt
requests
python-dotenv
neo4j
prettytable